+++ source/test-java/org/alfresco/repo/publishing/flickr/FlickrTest.java
@@ -25,6 +25,7 @@ import java.util.Map;
 
 import org.alfresco.model.ContentModel;
 import org.alfresco.repo.content.MimetypeMap;
+import org.alfresco.repo.node.archive.NodeArchiveService;
 import org.alfresco.repo.publishing.Environment;
 import org.alfresco.repo.publishing.PublishingModel;
 import org.alfresco.repo.publishing.PublishingQueueImpl;
@@ -39,6 +40,7 @@ import org.alfresco.service.cmr.repository.ContentService;
 import org.alfresco.service.cmr.repository.ContentWriter;
 import org.alfresco.service.cmr.repository.NodeRef;
 import org.alfresco.service.cmr.repository.NodeService;
+import org.alfresco.service.cmr.site.SiteInfo;
 import org.alfresco.service.cmr.site.SiteService;
 import org.alfresco.service.cmr.site.SiteVisibility;
 import org.alfresco.service.namespace.NamespaceService;
@@ -67,6 +69,7 @@ public class FlickrTest extends BaseSpringTest
     protected PublishingQueueImpl queue;
     protected Environment environment;
     protected NodeRef docLib;
+    protected NodeArchiveService nodeArchiveService;
 
     private ChannelService channelService;
     
@@ -76,6 +79,7 @@ public class FlickrTest extends BaseSpringTest
     {
         serviceRegistry = (ServiceRegistry) getApplicationContext().getBean("ServiceRegistry");
         channelService = (ChannelService) getApplicationContext().getBean("channelService"); 
+        nodeArchiveService = (NodeArchiveService) getApplicationContext().getBean("nodeArchiveService");
         AuthenticationUtil.setFullyAuthenticatedUser(AuthenticationUtil.getAdminUserName());
         siteService = serviceRegistry.getSiteService();
         fileFolderService = serviceRegistry.getFileFolderService();
@@ -90,7 +94,12 @@ public class FlickrTest extends BaseSpringTest
 
     public void onTearDown()
     {
-        siteService.deleteSite(siteId);
+        SiteInfo siteInfo = siteService.getSite(siteId);
+        if (siteInfo != null)
+        {
+            siteService.deleteSite(siteId);
+            nodeArchiveService.purgeArchivedNode(nodeArchiveService.getArchivedNode(siteInfo.getNodeRef()));
+        }
     }
     
     public void testBlank()
