+++ source/test-java/org/alfresco/repo/publishing/youtube/YouTubeTest.java
@@ -24,6 +24,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import org.alfresco.model.ContentModel;
+import org.alfresco.repo.node.archive.NodeArchiveService;
 import org.alfresco.repo.publishing.Environment;
 import org.alfresco.repo.publishing.PublishingModel;
 import org.alfresco.repo.publishing.PublishingQueueImpl;
@@ -38,6 +39,7 @@ import org.alfresco.service.cmr.repository.ContentService;
 import org.alfresco.service.cmr.repository.ContentWriter;
 import org.alfresco.service.cmr.repository.NodeRef;
 import org.alfresco.service.cmr.repository.NodeService;
+import org.alfresco.service.cmr.site.SiteInfo;
 import org.alfresco.service.cmr.site.SiteService;
 import org.alfresco.service.cmr.site.SiteVisibility;
 import org.alfresco.service.namespace.NamespaceService;
@@ -66,6 +68,7 @@ public class YouTubeTest extends BaseSpringTest
     protected PublishingQueueImpl queue;
     protected Environment environment;
     protected NodeRef docLib;
+    protected NodeArchiveService nodeArchiveService;
 
     private ChannelService channelService;
     
@@ -75,6 +78,7 @@ public class YouTubeTest extends BaseSpringTest
     {
         serviceRegistry = (ServiceRegistry) getApplicationContext().getBean("ServiceRegistry");
         channelService = (ChannelService) getApplicationContext().getBean("channelService"); 
+        nodeArchiveService = (NodeArchiveService) getApplicationContext().getBean("nodeArchiveService");
         AuthenticationUtil.setFullyAuthenticatedUser(AuthenticationUtil.getAdminUserName());
         siteService = serviceRegistry.getSiteService();
         fileFolderService = serviceRegistry.getFileFolderService();
@@ -89,7 +93,12 @@ public class YouTubeTest extends BaseSpringTest
 
     public void onTearDown()
     {
-        siteService.deleteSite(siteId);
+        SiteInfo siteInfo = siteService.getSite(siteId);
+        if (siteInfo != null)
+        {
+            siteService.deleteSite(siteId);
+            nodeArchiveService.purgeArchivedNode(nodeArchiveService.getArchivedNode(siteInfo.getNodeRef()));
+        }
     }
     
     public void testBlank()
